Update and move karma.conf.js
