--- {{cookiecutter.slug}}/frontend.angular/src/karma.conf.js
@@ -1,51 +1,0 @@
-// Karma configuration file, see link for more information
-// https://karma-runner.github.io/1.0/config/configuration-file.html
-
-module.exports = function (config) {
-    let configuration = {
-        basePath: '',
-        frameworks: ['jasmine', '@angular-devkit/build-angular'],
-        plugins: [
-            require('karma-jasmine'),
-            require('karma-chrome-launcher'),
-            require('karma-jasmine-html-reporter'),
-            require('karma-coverage-istanbul-reporter'),
-            require('@angular-devkit/build-angular/plugins/karma')
-        ],
-        client: {
-            clearContext: false // leave Jasmine Spec Runner output visible in browser
-        },
-        coverageIstanbulReporter: {
-            dir: require('path').join(__dirname, '../coverage/{{cookiecutter.slug}}'),
-            reports: ['html', 'lcovonly', 'text-summary'],
-            fixWebpackSourcePaths: true
-        },
-
-        reporters: ['progress', 'kjhtml'],
-        port: 9876,
-        colors: true,
-        logLevel: config.LOG_INFO,
-        autoWatch: true,
-        browsers: ['ChromeHeadless'],
-        customLaunchers: {
-            ChromeHeadless: {
-                base: 'Chrome',
-                flags: [
-                    '--headless',
-                    '--disable-gpu',
-                    '--no-sandbox',
-                    '--remote-debugging-port=9222',
-                ]
-            }
-        },
-        singleRun: false,
-        restartOnFileChange: true
-    };
-
-    if (process.env.CI) {
-        configuration.browsers = ['ChromeHeadless'];
-        configuration.singleRun = true;
-    }
-
-    config.set(configuration);
-};
